requirements.txt
# seaborn
# mediapipe
opencv-python~=4.8.0.76
numpy~=1.24
matplotlib~=3.5.3
pandas~=1.3.5
pillow~=9.5.0
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
